Map now loads from firebase automatically. - Added the map css and some required javascript files (map.css, circle.js, line.js, map.js, rectangle.js, table.js)
What works?
- Automatically map loading
- User can't modify the map

What doesn't work?
- Getting a group from firebase, but instructions to get a group are posted. See below for more.

The method CheckIfTable(mouseX, mouseY) in map.js file verifies that the user has clicked on a table (square with number).
TO DO: Add a method which receives the group related to the groupnumber of the table you clicked on from firebase and visualize it.

Visualization can be done in an alert as it's just to prove that you can get a list of groupmembers related to the clicked table.